Move forcing and example parameter sets to plugins (and lisflood config)
