+++ .gitignore
@@ -27,3 +27,5 @@ docs/apidocs/
 
 docs/examples/parameter-sets
 docs/examples/ewatercycle.yaml
+
+__pycache__
