+++ src/ewatercycle/plugins/lisflood/example_parameterset.py
@@ -1,6 +1,6 @@
 from typing import Iterable
 
-from ._example import ExampleParameterSet
+from ...parameter_sets._example import ExampleParameterSet
 
 
 def example_parameter_sets() -> Iterable[ExampleParameterSet]:
