+++ src/ewatercycle/plugins/lisflood/model.py
@@ -10,12 +10,12 @@ import xarray as xr
 from cftime import num2date
 
 from ewatercycle import CFG
-from ewatercycle.config._lisflood_versions import version_images
 from ewatercycle.container import start_container
-from ewatercycle.forcing._lisflood import LisfloodForcing
 from ewatercycle.models.abstract import AbstractModel
 from ewatercycle.parameter_sets import ParameterSet
 from ewatercycle.parametersetdb.config import XmlConfig
+from ewatercycle.plugins.lisflood._lisflood_versions import version_images
+from ewatercycle.plugins.lisflood.forcing import LisfloodForcing
 from ewatercycle.util import find_closest_point, get_time, to_absolute_path
 
 logger = logging.getLogger(__name__)
