+++ src/ewatercycle/plugins/marrmot/forcing.py
@@ -5,9 +5,9 @@ from typing import Literal, Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ..util import get_time, to_absolute_path
-from ._default import DefaultForcing, _session
-from .datasets import DATASETS
+from ewatercycle.forcing._default import DefaultForcing, _session
+from ewatercycle.forcing.datasets import DATASETS
+from ewatercycle.util import get_time, to_absolute_path
 
 
 class MarrmotForcing(DefaultForcing):
