+++ src/ewatercycle/plugins/wflow/example_parameterset.py
@@ -1,6 +1,6 @@
 from typing import Iterable
 
-from ._example import ExampleParameterSet
+from ewatercycle.parameter_sets._example import ExampleParameterSet
 
 
 def example_parameter_sets() -> Iterable[ExampleParameterSet]:
