+++ tests/forcing/test_hype.py
@@ -7,7 +7,7 @@ from esmvalcore.experimental.recipe_output import OutputFile
 
 from ewatercycle.forcing import generate, load
 from ewatercycle.forcing._default import FORCING_YAML
-from ewatercycle.forcing._hype import HypeForcing
+from ewatercycle.plugins.hype.forcing import HypeForcing
 
 
 def test_plot():
