+++ tests/forcing/test_lisflood.py
@@ -8,7 +8,7 @@ from esmvalcore.experimental.recipe_output import DataFile
 
 from ewatercycle.forcing import generate, load
 from ewatercycle.forcing._default import FORCING_YAML
-from ewatercycle.forcing._lisflood import LisfloodForcing
+from ewatercycle.plugins.lisflood.forcing import LisfloodForcing
 
 
 def test_plot():
