+++ tests/forcing/test_wflow.py
@@ -9,7 +9,7 @@ from esmvalcore.experimental.recipe_output import DataFile
 
 from ewatercycle.forcing import generate, load
 from ewatercycle.forcing._default import FORCING_YAML
-from ewatercycle.forcing._wflow import WflowForcing
+from ewatercycle.plugins.wflow.forcing import WflowForcing
 
 
 @pytest.fixture
